CHANGED: AddNodeVariableDependencies is checked at run time
CHANGED: option -check_dependencies_at_build_time

Pbsfile md5 is ignored and only the digest content is checked!! Be careful this put the burden
of a correct digest on the rule developer.

this also means that cahanges to builder defined inline, subs or shell commands, are not
detected, subs can be defined in PbsUse'd files and their md5 is automatically added to the
node digest, but also trigger all the nodes.

This is development feature to not be used in production till it is polished

need to be added:
	detect changes in the rule, and it's elements, per node
	make PbsUse add depenedency to only those nodes that use something from it