ADDED: implemented PACKAGE_CONFIG_NO_INHERITANCE
PACKAGE_CONFIG_NO_INHERITANCE is  field that can be added to a subpbs declaration.
It is closely related to PACKAGE_CONFIG; they both control the configuration of
subpbs.

if you set  PACKAGE_CONFIG_NO_INHERITANCE => 1, the current config is not inherited
by the subpbs. TAGET_PATH is still set and any configuration in PACKAGE_CONFIG is also used.
This allows control from the parent side. There is no control mechanism from the subpbs
side, that would make no sense in a hierarchical system. It is possible from the subpbs, if
really wanted, to verify what configuration is available by calling GetConfigKeys and act
upon it.

CHANGED: --display_depend_start to dislpay_depend_end, the start is always show.