ADDED: --trigger family --gtg_node_start and changed exit code
--trigger family:
-----------------
A set of utilities to check what effect node triggering has on the dependency graph.

All the ooptions are debug options and stop the build.

--trigger_none, stops nodes that would trigged from triggering.

--triger=regex, makes a node trigger, multiple --trigger can be given

--triger_list=file, a list of regexes as if they were given through --trigger

Note that source node can't be triggere as it would mean rebuilding them and the build
step would fail as they have no builders, but pbs triggers their dependent
-----------------

--gtg_node_start=node_name, let's one point at the top node of the graph

exit code when DEBUG flags are set reflects an error to avoid people thinking the
build succeeded; it only succeeded providing information, not a system

Change the sort of nodes in tree to alphabetic again to reduced sort time and make it
easier to find nodes. Also in preparation of parallel depend where insertion time
means nothing.